+++ test/shell/test-smoke.sh
@@ -20,9 +20,9 @@ case ${1-0} in
       run_test "$0" 1 2
       exit
       ;;
-  1 ) echo "# 1: composer 2 must not have written lock"
+  1 ) echo "# 1: composer 1 must not have written lock"
       cd "$PROJECT_DIR"
-      if grep -cq '"plugin-api-version": "2\..*"' composer.lock; then exit 1; fi;
+      if grep -cq '"plugin-api-version": "1\..*"' composer.lock; then exit 1; fi;
       exit
       ;;
   2 ) echo "# 2: composer which"